+++ core/data/src/main/java/com/coldblue/data/repository/todo/MandaTodoRepositoryImpl.kt
@@ -1,5 +1,6 @@
 package com.coldblue.data.repository.todo
 
+import com.coldblue.data.alarm.AlarmScheduler
 import com.coldblue.data.mapper.MandaTodoMapper.asDomain
 import com.coldblue.data.mapper.MandaTodoMapper.asEntity
 import com.coldblue.data.mapper.MandaTodoMapper.asNetworkModel
@@ -8,15 +9,19 @@ import com.coldblue.data.mapper.asEntity
 import com.coldblue.data.mapper.asNetworkModel
 import com.coldblue.data.mapper.asSyncedEntity
 import com.coldblue.data.sync.SyncHelper
+import com.coldblue.data.util.isPassed
 import com.coldblue.database.dao.MandaTodoDao
 import com.coldblue.datastore.UpdateTimeDataSource
+import com.coldblue.model.AlarmItem
 import com.coldblue.model.MandaTodo
+import com.coldblue.model.Todo
 import com.coldblue.network.datasource.MandaTodoDataSource
 import com.orhanobut.logger.Logger
 import kotlinx.coroutines.flow.Flow
 import kotlinx.coroutines.flow.first
 import kotlinx.coroutines.flow.map
 import java.time.LocalDate
+import java.time.LocalDateTime
 import javax.inject.Inject
 
 class MandaTodoRepositoryImpl @Inject constructor(
@@ -24,9 +29,11 @@ class MandaTodoRepositoryImpl @Inject constructor(
     private val mandaTodoDataSource: MandaTodoDataSource,
     private val syncHelper: SyncHelper,
     private val updateTimeDataSource: UpdateTimeDataSource,
+    private val alarmScheduler: AlarmScheduler,
 ) : MandaTodoRepository {
     override suspend fun upsertMandaTodo(mandaTodo: MandaTodo) {
         mandaTodoDao.upsertMandaTodo(mandaTodo.asEntity())
+        mandaTodo.syncAlarm()
         syncHelper.syncWrite()
     }
 
@@ -63,6 +70,25 @@ class MandaTodoRepositoryImpl @Inject constructor(
             Logger.e("${e.message}")
             return false
         }
+    }
+
+    private suspend fun MandaTodo.syncAlarm() {
+        // 시간 null 체크
+        if (time == null) {
+            alarmScheduler.cancel(id)
+            return
+        }
+        // date, time 추가 시점이 과거인지 체크
+        if (LocalDateTime.of(date, time).isPassed()) {
+            alarmScheduler.cancel(id)
+            return
+        }
+        // 삭제, 완료했는지 체크
+        if (isDel or isDone) {
+            alarmScheduler.cancel(id)
+            return
+        }
+        alarmScheduler.add(AlarmItem(LocalDateTime.of(date, time), title, id))
     }
 
 }
