+++ feature/todo/src/main/java/com/coldblue/todo/MandaTodoList.kt
@@ -14,7 +14,6 @@ import androidx.compose.foundation.layout.offset
 import androidx.compose.foundation.layout.padding
 import androidx.compose.foundation.layout.size
 import androidx.compose.foundation.layout.width
-import androidx.compose.foundation.layout.wrapContentHeight
 import androidx.compose.foundation.lazy.LazyColumn
 import androidx.compose.foundation.lazy.LazyRow
 import androidx.compose.foundation.lazy.items
@@ -25,24 +24,11 @@ import androidx.compose.foundation.text.BasicTextField
 import androidx.compose.material.icons.Icons
 import androidx.compose.material.icons.automirrored.filled.KeyboardArrowRight
 import androidx.compose.material.icons.filled.KeyboardArrowDown
-import androidx.compose.material.icons.filled.KeyboardArrowRight
 import androidx.compose.material.icons.rounded.Clear
-import androidx.compose.material3.Button
-import androidx.compose.material3.ButtonDefaults
-import androidx.compose.material3.Card
-import androidx.compose.material3.DatePicker
-import androidx.compose.material3.DatePickerDefaults
-import androidx.compose.material3.DatePickerDialog
-import androidx.compose.material3.DisplayMode
-import androidx.compose.material3.ExperimentalMaterial3Api
 import androidx.compose.material3.Icon
 import androidx.compose.material3.IconButton
-import androidx.compose.material3.SelectableDates
 import androidx.compose.material3.Surface
 import androidx.compose.material3.Text
-import androidx.compose.material3.TimePicker
-import androidx.compose.material3.rememberDatePickerState
-import androidx.compose.material3.rememberTimePickerState
 import androidx.compose.runtime.Composable
 import androidx.compose.runtime.derivedStateOf
 import androidx.compose.runtime.getValue
@@ -61,10 +47,8 @@ import androidx.compose.ui.text.style.TextDecoration
 import androidx.compose.ui.text.style.TextOverflow
 import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.dp
-import androidx.compose.ui.window.Dialog
 import com.coldblue.data.util.toMillis
 import com.coldblue.designsystem.IconPack
-import com.coldblue.designsystem.component.HMButton
 import com.coldblue.designsystem.iconpack.todo.AddSquare
 import com.coldblue.designsystem.iconpack.todo.Alarm
 import com.coldblue.designsystem.iconpack.todo.Calendar
@@ -77,12 +61,10 @@ import com.coldblue.model.MandaTodo
 import com.coldblue.model.MyDate
 import com.coldblue.model.MyTime
 import com.coldblue.model.ToggleInfo
-import java.text.SimpleDateFormat
 import java.time.LocalDate
 import java.time.LocalDateTime
 import java.time.LocalTime
 import java.time.format.DateTimeFormatter
-import java.util.Date
 import java.util.Locale
 
 fun getDisplayTime(time: LocalTime?): String {
@@ -500,293 +482,6 @@ fun TodoInput(
 }
 
 @Composable
-fun TodoDialog(
-    onClickCancel: () -> Unit,
-    mandaTodo: MandaTodo,
-    upsertMandaTodo: (MandaTodo) -> Unit,
-) {
-    var timePickerState by remember { mutableStateOf(false) }
-    var datePickerState by remember { mutableStateOf(false) }
-
-    var text by remember { mutableStateOf(mandaTodo.title) }
-
-
-    if (datePickerState) {
-        CustomDatePickerDialog(
-            LocalDateTime.of(mandaTodo.date, LocalTime.now()).toMillis(),
-            { datePickerState = false },
-            {
-                datePickerState = false
-                upsertMandaTodo(
-                    mandaTodo.copy(
-                        date = LocalDate.parse(
-                            it,
-                            DateTimeFormatter.BASIC_ISO_DATE
-                        )
-                    )
-                )
-            }
-        )
-    }
-    if (timePickerState) {
-        CustomTimePickerDialog(
-            if (mandaTodo.time != null) mandaTodo.time?.hour else LocalTime.now().hour,
-            if (mandaTodo.time != null) mandaTodo.time?.minute else LocalTime.now().minute,
-            { timePickerState = false },
-            { h, m ->
-                timePickerState = false
-                upsertMandaTodo(mandaTodo.copy(time = LocalTime.of(h, m)))
-            }
-        )
-    }
-    Dialog(
-        onDismissRequest = { onClickCancel() },
-    ) {
-        Card(
-            shape = RoundedCornerShape(8.dp),
-        ) {
-            Column(
-                modifier = Modifier
-                    .background(HMColor.Background)
-                    .padding(16.dp)
-            ) {
-                LazyColumn {
-                    item {
-                        BasicTextField(
-                            textStyle = HmStyle.text16.copy(fontWeight = FontWeight.Bold),
-                            modifier = Modifier
-                                .fillMaxWidth()
-                                .padding(top = 16.dp),
-                            value = text, onValueChange = { text = it },
-                        )
-                    }
-                }
-                val isDateSet = mandaTodo.date == null
-                Row(
-                    modifier = Modifier
-                        .fillMaxWidth()
-                        .clickable { datePickerState = true }
-                        .padding(vertical = 12.dp),
-                    verticalAlignment = Alignment.CenterVertically
-                ) {
-                    val color =
-                        if (isDateSet) HMColor.DarkGray else HMColor.Primary
-                    val dateText =
-                        if (isDateSet) "날짜 추가" else mandaTodo.date.toString()
-                    val textColor =
-                        if (isDateSet) HMColor.DarkGray else HMColor.Text
-                    Icon(
-                        imageVector = IconPack.Calendar,
-                        contentDescription = "",
-                        tint = color,
-                    )
-                    Spacer(modifier = Modifier.width(20.dp))
-                    Text(dateText, color = textColor)
-                }
-
-                val isTimeSet = mandaTodo.time != null
-                Row(
-                    modifier = Modifier
-                        .fillMaxWidth()
-                        .clickable { timePickerState = true }
-                        .padding(vertical = 12.dp),
-                    verticalAlignment = Alignment.CenterVertically,
-                    horizontalArrangement = Arrangement.SpaceBetween
-                ) {
-                    val color =
-                        if (isTimeSet) HMColor.Primary else HMColor.DarkGray
-                    val timeText =
-                        if (isTimeSet) getDisplayTime(mandaTodo.time) else "알림 추가"
-                    val textColor =
-                        if (isTimeSet) HMColor.Text else HMColor.DarkGray
-                    Row {
-                        Icon(
-                            imageVector = IconPack.Alarm,
-                            contentDescription = "",
-                            tint = color,
-                        )
-                        Spacer(modifier = Modifier.width(20.dp))
-                        Text(timeText, color = textColor)
-                    }
-
-                    if (isTimeSet) {
-                        Icon(
-                            modifier = Modifier
-                                .size(20.dp)
-                                .clip(CircleShape)
-                                .clickable {
-                                    upsertMandaTodo(mandaTodo.copy(time = null))
-                                },
-                            imageVector = Icons.Rounded.Clear,
-                            contentDescription = "",
-                            tint = HMColor.Primary,
-                        )
-                    }
-                }
-
-                Row(modifier = Modifier.padding(top = 12.dp)) {
-                    Button(
-                        modifier = Modifier
-                            .padding(end = 5.dp)
-                            .height(50.dp)
-                            .weight(1f),
-                        colors = ButtonDefaults.buttonColors(containerColor = HMColor.Gray),
-                        shape = RoundedCornerShape(10.dp),
-                        onClick = {
-                            onClickCancel()
-                            upsertMandaTodo(mandaTodo.copy(isDel = true))
-                        }
-                    ) {
-                        Text(
-                            text = "삭제",
-                            style = HmStyle.text16,
-                            color = HMColor.Primary,
-                            fontWeight = FontWeight.Bold
-                        )
-                    }
-                    HMButton(
-                        "저장",
-                        text.isNotBlank(),
-                        modifier = Modifier.fillMaxWidth(0.5f),
-                        onClick = {
-                            onClickCancel()
-                            upsertMandaTodo(mandaTodo.copy(title = text))
-                        })
-                }
-
-            }
-        }
-    }
-
-}
-
-@Preview
-@Composable
-fun TodoDialogPreview() {
-    TodoDialog({}, MandaTodo("내용입니ㅏ 내용입니ㅏ 내용입니ㅏ 내용입니ㅏ내용입니ㅏ", mandaIndex = 2), {})
-}
-
-@OptIn(ExperimentalMaterial3Api::class)
-@Composable
-fun CustomDatePickerDialog(
-    selectedDate: Long,
-    onClickCancel: () -> Unit,
-    onClickConfirm: (yyyyMMdd: String) -> Unit
-) {
-    DatePickerDialog(
-        onDismissRequest = { onClickCancel() },
-        confirmButton = {},
-        colors = DatePickerDefaults.colors(
-            containerColor = Color.White
-        ),
-        shape = RoundedCornerShape(6.dp)
-    ) {
-        val datePickerState = rememberDatePickerState(
-            yearRange = LocalDate.now().year..LocalDate.now().year + 1,
-            initialDisplayMode = DisplayMode.Picker,
-            initialSelectedDateMillis = selectedDate,
-            selectableDates = object : SelectableDates {
-                override fun isSelectableDate(utcTimeMillis: Long): Boolean {
-                    return true
-                }
-            })
-
-        DatePicker(
-            state = datePickerState,
-        )
-
-        Row(
-            modifier = Modifier.fillMaxWidth(),
-            horizontalArrangement = Arrangement.End,
-        ) {
-            Button(onClick = {
-                onClickCancel()
-            }) {
-                Text(text = "취소")
-            }
-            Spacer(modifier = Modifier.width(5.dp))
-
-            Button(onClick = {
-                datePickerState.selectedDateMillis?.let { selectedDateMillis ->
-                    val yyyyMMdd = SimpleDateFormat(
-                        "yyyyMMdd",
-                        Locale.getDefault()
-                    ).format(Date(selectedDateMillis))
-
-                    onClickConfirm(yyyyMMdd)
-                }
-            }) {
-                Text(text = "확인")
-            }
-        }
-    }
-}
-
-@OptIn(ExperimentalMaterial3Api::class)
-@Composable
-fun CustomTimePickerDialog(
-    selectedHour: Int?,
-    selectedMinute: Int?,
-    onClickCancel: () -> Unit,
-    onClickConfirm: (hour: Int, minute: Int) -> Unit
-) {
-    val timePickerState = rememberTimePickerState(
-        initialHour = selectedHour ?: 0,
-        initialMinute = selectedMinute ?: 0,
-        is24Hour = false
-    )
-    Dialog(
-        onDismissRequest = { onClickCancel() },
-    ) {
-        Card(
-            shape = RoundedCornerShape(8.dp),
-        ) {
-            Column(
-                modifier = Modifier
-                    .width(300.dp)
-                    .wrapContentHeight()
-                    .background(
-                        color = Color.White,
-                    )
-                    .padding(16.dp),
-                horizontalAlignment = Alignment.CenterHorizontally,
-            ) {
-
-                Text(text = "시간을 선택해주세요.")
-
-                Spacer(modifier = Modifier.width(15.dp))
-
-                TimePicker(
-                    state = timePickerState,
-                    modifier = Modifier.padding(top = 10.dp),
-                )
-
-                Row(
-                    modifier = Modifier.fillMaxWidth(),
-                    horizontalArrangement = Arrangement.Center,
-                ) {
-                    Button(onClick = {
-                        onClickCancel()
-                    }) {
-                        Text(text = "취소")
-                    }
-
-                    Spacer(modifier = Modifier.width(5.dp))
-
-                    Button(onClick = {
-                        val hour = timePickerState.hour
-                        val minute = timePickerState.minute
-                        onClickConfirm(hour, minute)
-                    }) {
-                        Text(text = "확인")
-                    }
-                }
-            }
-        }
-    }
-}
-
-@Composable
 fun TodoRangeSelector(todoRange: DateRange, changeRange: (DateRange) -> Unit) {
     val dateRangeButtons = remember {
         mutableStateListOf(
@@ -854,5 +549,4 @@ fun MandaTodoItemPreview() {
         ),
         3, 3, {}, {},
     )
-
 }
